Function that searches for a value in an array of integers using the Binary search algorithm

+++ 0x1E-search_algorithms/1-binary.c
@@ -25,7 +25,7 @@ while (test == 0 && start <= end)
 mil = (start + end) / 2;
 if (array[mil] == value)
 {
-return (mil);
+return ((int) mil);
 }
 else if (value > array[mil])
 {
